+++ popkorn/src/commonMain/kotlin/cc/popkorn/core/exceptions/PopKornNotInitializedException.kt
@@ -1,0 +1,3 @@
+package cc.popkorn.core.exceptions
+
+class PopKornNotInitializedException : RuntimeException("You must execute PopKornCompatKt.setup(...) before using this library")