+++ popkorn/src/nativeMain/kotlin/cc/popkorn/Platform.kt
@@ -1,6 +1,7 @@
 package cc.popkorn
 
 import cc.popkorn.core.Injector
+import cc.popkorn.core.exceptions.NonExistingClassException
 import cc.popkorn.core.exceptions.PopKornNotInitializedException
 import cc.popkorn.mapping.Mapping
 import cc.popkorn.pools.MappingProviderPool
@@ -11,37 +12,28 @@ import kotlin.reflect.KClass
 
 
 /**
- * Implementation for JS of the methods/classes that are Platform-dependent
+ * Implementation for Native of the methods/classes that are Platform-dependent
  *
  * @author Pau Corbella
  * @since 1.6.0
  */
 
 
-actual class WeakReference<T : Any> actual constructor(referred: T) {
-    private var pointer: T? = referred
+actual typealias WeakReference<T> = kotlin.native.ref.WeakReference<T>
 
-    actual fun clear() {
-        pointer = null
-    }
-
-    actual fun get() = pointer
-
-}
-
-internal actual fun <T : Any> KClass<T>.getName() = js.name
+internal actual fun <T : Any> KClass<T>.getName() = this.qualifiedName ?: throw NonExistingClassException(this)
 
 internal actual fun <T : Any> KClass<T>.needsResolver(resolverPool: ResolverPool) = resolverPool.isPresent(this)
 
-internal actual fun createDefaultInjector() = Injector(jsResolverPool(), jsProviderPool())
+internal actual fun createDefaultInjector() = Injector(nativeResolverPool(), nativeProviderPool())
 
 
-private fun jsResolverPool(): ResolverPool {
+private fun nativeResolverPool(): ResolverPool {
     if (!::iResolverMappings.isInitialized || !::iProviderMappings.isInitialized) throw PopKornNotInitializedException()
     return MappingResolverPool(iProviderMappings)
 }
 
-private fun jsProviderPool(): ProviderPool {
+private fun nativeProviderPool(): ProviderPool {
     if (!::iResolverMappings.isInitialized || !::iProviderMappings.isInitialized) throw PopKornNotInitializedException()
     return MappingProviderPool(iProviderMappings)
 }